\name{deldir-internal}
\alias{[.tile.list}
\alias{[.triang.list}
\alias{acw}
\alias{binsrtR}
\alias{dumpts}
\alias{get.cnrind}
\alias{getCol}
\alias{mid.in}
\alias{mnndR}
\alias{prelimtlist}
\alias{tilePerim0}
\title{Internal deldir functions}
\description{
  Internal deldir functions.
}
\usage{
\method{[}{tile.list}(x,i,\dots)
\method{[}{triang.list}(x,i,\dots)
acw(xxx)
dumpts(x,y,dpl,rw)
binsrtR(x,y,rw)
get.cnrind(x,y,rw)
getCol(x,warn=FALSE)
mid.in(x,y,rx,ry)
mnndR(x,y)
prelimtlist(object)
tilePerim0(object,inclbdry=TRUE)
}
\details{
  These functions are auxiliary and are not intended to be called by
  the user.
}
\keyword{internal}
