\name{deldir-internal}
\alias{dumpts}
\alias{ind.dup}
\alias{mnnd}
\title{Internal deldir functions}
\description{
  Internal deldir functions.
}
\usage{
dumpts(x,y,dpl,rw)
ind.dup(x,y,rw=NULL,frac=0.0001)
mnnd(x,y)
}
\details{
  These functions are auxilliary and are not intended to be called by
  the user.
}
\keyword{internal}
