\name{deldir-internal}
\alias{dumpts}
\alias{ind.dup}
\alias{mid.in}
\alias{mnnd}
\alias{get.cnrind}
\alias{acw}
\title{Internal deldir functions}
\description{
  Internal deldir functions.
}
\usage{
dumpts(x,y,dpl,rw)
ind.dup(x,y,rw=NULL,frac=0.0001)
mid.in(x,y,rx,ry)
mnnd(x,y)
get.cnrind(x,y,rw)
acw(xxx)
}
\details{
  These functions are auxilliary and are not intended to be called by
  the user.
}
\keyword{internal}
